requirements.txt
pdfplumber
beautifulsoup4
requests
scikit-learn
chardet
numpy
transformers
torch
langchain_groq
python-dotenv
langchain
langchain_community
langchain_core
langchainhub
langchain_google_genai
langchain_text_splitters
langchain-objectbox
faiss-cpu
sentence-transformers